test setup variation
